Reject invalid user input and reprompt
Human player is notified of their error and reprompted if they enter
anything besides rock, paper, or scissors, including just hitting "okay"
without typing anything. Their choice is not case-sensitive.

Only remaining bug is that "cancel" still aborts the game with an
error message rather than trapping the user inside the game loop
forever or alternatively leading to an elegant option such as
starting a new game or confirming that they are ready to move on with
their life, etc.