Add table that dynamically displays round history
A table at the bottom of the page now adds a line as each round is
finished that displays the round #, human's choice of object, computer's
choice of object, and round winner.

This goes beyond TOP assignment but it's a fun addition and seems like a
useful feature to know how to build.