Add colors and basic formatting
Page items are reasonably positioned and there is a sort of orange-red,
olive green, and dusty turquoise color scheme going on that I like.

Definitely room for more to be done but it's actually pretty cute and
looks great on mobile. For some reason when it loads in my browser most
of the viewport is the main divs stretching out in long horizontal bands
with the whole thing gathered toward the top of the screen while the
other content is tiny in the middle of the screen -- and the same is
true for iPad previews. Need to figure out what's going on with that.
Everything else style-wise at this point is just tweaking and playing
around with stuff.