Improve timing of announcement and update events
When the user selects an object, relevant information rolls out in a
logical fashion over several seconds, with time to understand what's
going on before it resets:

the human object displays
the robot object displays
the winner is announced
the score updates
the documentation updates

Then after a couple seconds the objects and winner announcement
disappear and the user is prompted to choose another object, or the
GAME OVER announcement appears if someone has 5 points.

Almost everything in the game works really well now (including new game
starting, which has not broken again so far)...as long as the user plays
along and waits for all the events to go down. If the user keeps hitting
buttons in rapid succession weird stuff can happen, including a player
accruing more than 5 points before the GAME OVER announcement
materializes or the displaying of the object choices generally getting
wonky until and if the program can catch up.

Which means the one remaining task (????) is to render the rock-paper-
scissors buttons unclickable until all the requisite events have taken
place each round.